The compiling file finally compiles an executable
The original placeholder bin has been replaced with the full path of the
where the glfw3.dll is being held. Gdi32 was added here as well, without
it, glfwInit sends a bunch of errors. The executable didn't need to be
taken from the src file since it compiled to the root dir, so it was
just sent to the bin normally.